% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_abnormal_subjects.R
\name{check_abnormal_subjects}
\alias{check_abnormal_subjects}
\title{Sensitivity check}
\usage{
check_abnormal_subjects(values)
}
\description{
In iterative weighted least square methods,
each data point is transformed into a corresponding pseudo-response with
a corresponding pseudo-variance.
These transformations would inevitably point to some abnormal subjects
whose pseudo-response or pseudo-variance is an abnormal value,
such as Inf/-Inf and NaN(?).
We take these subjects with abnormal pseudo-response/pseudo-variance
as \dQuote{abnormal subjects}.
\code{check_abnormal_subjects()} return the indices of abnormal
subjects (if any).
}
