% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_colstats}
\alias{compute_colstats}
\title{Auxilary computation functions}
\usage{
compute_colstats(X, scale = TRUE)
}
\description{
This file defines the auxiliary computation functions.
}
\details{
\code{compute_colstats} computes the mean and standard deviation of each
column.

\param{scale} Boolean, indicator to perform standardization or not.
Call \code{scale = TRUE} to compute the column-wise means and standard
deviations (sd). If \code{scale = FALSE}, the output mean and sd vectors are,
respectively, all-0 and all-1 vector of length equal the number of columns of
\eqn{X}.

It returns a list with two attributes:
\code{out$cm} is a p-vector of column-wise means of X if
\code{scale=TRUE}, a p-dim vector of zeros otherwise;
\code{out$csd} is a p-dim vector of column-wise standard
deviations of X if \code{scale=TRUE}, a p-dim vector of ones otherwise;

\code{compute_Xb} offers the matrix multiplication of
an (n x p) matrix \code{X} and a (p x 1) array/matrix {b}.
The output is a (n x 1) matrix.
}
\keyword{internal}
