% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic.R
\name{compute_logw2_logistic}
\alias{compute_logw2_logistic}
\title{\code{compute_logw2_logistic} computes the log-pseudo-variance
for each observation}
\usage{
compute_logw2_logistic(eta)
}
\description{
\code{compute_logw2_logistic} computes the log-pseudo-variance
for each observation
}
\keyword{internal}
