% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_poisson.R
\name{compute_logw2_poisson}
\alias{compute_logw2_poisson}
\title{Functions related to Poisson regression model with log link}
\usage{
compute_logw2_poisson(eta)
}
\description{
This file defines functions related to Poisson regression model
with logistic link, including (log of) pseudo-variance and pseudo-response
that are needed to be calculated during iterative fitting of the
(approximated) weighted linear regression model, and the exact and
approximated log-likelihood.

The functions below output vectors of the same length of y

#' \code{compute_logw2_poisson} computes the log-pseudo-variance
}
\keyword{internal}
