% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic.R
\name{compute_psdresponse_logistic}
\alias{compute_psdresponse_logistic}
\title{\code{compute_psdresponse_logistic} computes the pseudo-response
for each observation}
\usage{
compute_psdresponse_logistic(eta, y)
}
\description{
\code{compute_psdresponse_logistic} computes the pseudo-response
for each observation
}
\keyword{internal}
