% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{KL_prior_v_post}
\alias{KL_prior_v_post}
\alias{get_objective}
\alias{clipped}
\title{Compute Evidence of Lower Bounds (ELBO)}
\usage{
KL_prior_v_post(sigma02, pie, alpha, mu, sigma12)

get_objective(X, y, gs, model, abn.rm = FALSE)

clipped(values, tol = 1e-16)
}
\arguments{
\item{sigma02}{a scalar, prior variance of coefficient}

\item{pie}{a p-dim vector of prior inclusion probabilities}

\item{alpha}{a p-dim vector of posterior inclusion probabilities}

\item{mu}{a p-dim vector of posterior means}

\item{sigma12}{a p-dim vector of posterior variances}
}
\value{
a value of KL-divergence between the prior and posterior (VA)
distributions of L WSER models.
}
\description{
This function is to compute the second part of ELBO of the lth WSER model
ELBO = Expected loglik + KL(g(b)||q(b)).

Compute the (estimated) ELBO for the overall model.

Clip the inputs to avoid feeding 0 into \code{log()}
}
\keyword{internal}
