% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsusie_coefficients.R
\name{gsusie_coefficients}
\alias{gsusie_coefficients}
\title{Summarize posterior estimations of regression coefficients}
\usage{
gsusie_coefficients(
  object,
  vars = NULL,
  decreasing = TRUE,
  top_n = 10,
  cred_int = TRUE,
  coverage = 0.95,
  digits = 4
)
}
\arguments{
\item{object}{a gsusie fit}

\item{vars}{an array of numerical indices or variable names
presented in the output. If \code{var_names = NULL},
by default the output contains coefficients of all variables
if \code{ncol(X)<= 10} or \code{top_n} variables with highest PIPs
if \code{ncol(X) > 10}.}

\item{decreasing}{logical, whether to list the variables
in decreasing order of PIP values.}

\item{top_n}{numeric, the number of coefficients to be displayed. By default
\code{top_n=min(10, length(object$pip))}.}

\item{cred_int}{logical, whether to return equal-tailed credible intervals}

\item{coverage}{numeric, between 0 and 1, the coverage probability
of credible intervals.}

\item{digits}{integer indicating the number of decimal places to be used.
if \code{digits == NULL}, the output will not be rounded.}
}
\description{
This function returns, by default, \code{top_n=10} coefficients with the
highest PIPs, sorted in decreasing order.
}
