% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsusie_plots.R
\name{gsusie_plot}
\alias{gsusie_plot}
\title{G-SuSiE Plots.}
\usage{
gsusie_plot(
  model,
  y,
  add_bar = FALSE,
  pos = NULL,
  b = NULL,
  max_cs = 400,
  add_legend = NULL,
  ...
)
}
\arguments{
\item{model}{A SuSiE or G-SuSiE fit, typically an output from
\code{\link{gsusie}} or one of its variants.
For \code{gsuse_plot},
the susie fit must have \code{model$z}(?), \code{model$PIP}, and may
include \code{model$sets}. \code{model} may also be a vector of
z-scores or PIPs.}

\item{y}{A string indicating what to plot: either \code{"z_original"} for
z-scores, \code{"z"} for z-score derived p-values on (base-10) log-scale,
\code{"PIP"} for posterior inclusion probabilities,
\code{"log10PIP"} for posterior inclusion probabiliities on the
(base-10) log-scale. For any other setting, the data are plotted as
is.}

\item{add_bar}{If \code{add_bar = TRUE}, add horizontal bar to
signals in credible interval.}

\item{pos}{This can be either be (1) a numeric vector of indices of
subset of variables to plot, or (2) a list with the following list
elements: \code{pos$attr}, \code{pos$start} and \code{pos$end},
where \code{pos$attr} is a character string of the name of index
variable in \code{model} object, and \code{pos$start} and
\code{pos$end} are boundaries of indices to plot. See the provided
examples.}

\item{b}{For simulated data, set \code{b = TRUE} to highlight
"true" effects (highlights in red).}

\item{max_cs}{The largest credible set to display, either based on
purity (set \code{max_cs} between 0 and 1), or based on size (set
\code{max_cs > 1}).}

\item{add_legend}{If \code{add_legend = TRUE}, add a legend to
annotate the size and purity of each CS discovered. It can also be
specified as location where legends should be added, e.g.,
\code{add_legend = "bottomright"} (default location is
\code{"topright"}).}

\item{\dots}{Additional arguments passed to
\code{\link[graphics]{plot}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
\code{gsusie_plot} produces a per-variable summary of
the G-SuSiE credible sets.
}
