% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_setup}
\alias{init_setup}
\title{Initialize a GSuSiE object using regression coefficients}
\usage{
init_setup(
  n,
  p,
  L,
  family,
  prior_inclusion_prob,
  coef_prior_variance,
  null_weight
)
}
\description{
Initialize a GSuSiE object using regression coefficients
}
\keyword{internal}
