% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_single_effect_regression.R
\name{neg.optimfunc.logscale}
\alias{neg.optimfunc.logscale}
\title{The following is the negative of the objective function}
\usage{
neg.optimfunc.logscale(lV, betahat, shat2, prior_inclusion_prob)
}
\description{
The following is the negative of the objective function
}
\keyword{internal}
