% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_single_effect_regression.R
\name{optimfunc.logscale}
\alias{optimfunc.logscale}
\title{The following is the log-scale of the optimization goal
as a function of prior variance V.}
\usage{
optimfunc.logscale(V, betahat, shat2, prior_inclusion_prob)
}
\description{
The following is the log-scale of the optimization goal
as a function of prior variance V.
}
\keyword{internal}
