% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_single_effect_regression.R
\name{optimize_prior_variance}
\alias{optimize_prior_variance}
\title{Estimate prior variance}
\usage{
optimize_prior_variance(
  optimize_V,
  betahat,
  shat2,
  prior_inclusion_prob,
  alpha = NULL,
  post_mean2 = NULL,
  V_init = NULL,
  check_null_threshold = 0
)
}
\description{
In this function, betahat represents the MLE,
and shat2 represents the corresponding variance.
}
\keyword{internal}
