% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gsusie_coefficients.R
\name{print_gsusie_coefficients}
\alias{print_gsusie_coefficients}
\title{Summarize posterior estimations of regression coefficients}
\usage{
print_gsusie_coefficients(
  object,
  subset_vars = NULL,
  decreasing = TRUE,
  top_n = 10,
  cred_int = TRUE,
  coverage = 0.95,
  digits = 4
)
}
\arguments{
\item{object}{a gsusie fit}

\item{subset_vars}{an array of numeric indices or variable names
presented in the output. If \code{var_names = NULL},
by default the output contains coefficients of all variables
if \code{ncol(X)<= 10} or \code{top_n} variables with highest PIPs
if \code{ncol(X) > 10}.}

\item{decreasing}{logical, whether to list the variables
in decreasing order of PIP values.}

\item{top_n}{numeric, the number of coefficients to be displayed. By default
\code{top_n=min(10, length(object$pip))}.}

\item{cred_int}{logical, whether to return equal-tailed credible intervals}

\item{coverage}{numeric, between 0 and 1, the coverage probability
of credible intervals.}

\item{digits}{integer indicating the number of decimal places to be used.
if \code{digits == NULL}, the output will not be rounded.}
}
\value{
The function outputs a data.frame with \code{top_n} variables with
the highest PIPs, or variables specified in \code{vars}.
Variables are sorted in decreasing order of PIP values.
The data.frame contains \emph{variable name}, \emph{PIP},
\emph{posterior mean} (\code{PostMean}),
\emph{posterior sd} (\code{PostSD}),
and, if \code{cred_int=TRUE}, \emph{credible interval}
(\code{CI_lower} and \code{CI_upper})
of each variable.
}
\description{
This function returns, by default, \code{top_n=10} coefficients with the
highest PIPs, sorted in decreasing order.
}
