% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_abnormal.R
\name{remove_abnormal}
\alias{remove_abnormal}
\title{Remove abnormal points from the current matrix/array}
\usage{
remove_abnormal(abn_index, object)
}
\arguments{
\item{abn_index}{an array of indices of abnormal points.}

\item{object}{a matrix or a vector that may contain abnormal points.
If \param{object} is a matrix, then \param{abn_index} indicates
the rows to be removed.}
}
\description{
Remove abnormal points from the current matrix/array
}
