% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_each_effect.R
\name{update_each_effect}
\alias{update_each_effect}
\title{Update each single effect at a time}
\usage{
update_each_effect(
  X,
  y,
  gs,
  model,
  estimate_prior_variance = FALSE,
  estimate_prior_method = "optim",
  check_null_threshold = 0,
  abnormal_proportion = 0.5,
  robust_estimation = FALSE,
  robust_method = "simple",
  simple_outlier_fraction = 0.01,
  simple_outlier_thres = NULL,
  robust_tuning_method = NULL
)
}
\arguments{
\item{X}{an n by p matrix of regression variables.}

\item{y}{an n vector of response variable.}

\item{gs}{a GSuSiE fit.}

\item{model}{a list containing functions required for
the specific generalized linear model.}

\item{estimate_prior_variance}{Boolean. If \code{TRUE}, the coefficient prior
variance is estimated and \code{coef_prior_variance} is then used as an
initial value for the optimization (if provided). If \code{FALSE}, the prior
variance for each of the \code{maxL} effects is fixed and determined by the
value supplied to \code{coef_prior_variance}.}

\item{estimate_prior_method}{The method used for estimating prior variance.
When \code{estimate_prior_method="simple"} is used, the likelihood at the
specified prior variance is compared to the likelihood at a variance of zero,
and the setting with the larger likelihood is retained.}

\item{check_null_threshold}{When the prior variance is estimated,
compare the estimated with the null, and set the prior variance to zero
unless the log-likelihood using the estimate is larger by this threshold
amount.}

\item{abnormal_proportion}{a value between 0 and 1.
If the number of detected abnormal subjects exceeds
\eqn{abnormal_proportion * nrow(X)}, stop fitting the model.}

\item{robust_estimation}{If \code{robust_estimation=TRUE},
robust estimation method is applied on the coefficient estimation.}

\item{robust_method}{If \code{robust_method= "simple"},
then the outliers defined by \code{simple_outlier_fraction} or
\code{simple_outlier_thres} are simply removed in each iteration.
If \code{robust_method="huber"}, Huber weights are additionally
multiplied to the pseudo-weights in each iteration.
If \code{robust_method = "bisquare"}, (Tukey's) bisquare weights are
additionally multipled to the pseudo-weights in each iteration.}

\item{simple_outlier_fraction}{a value between 0 and 1, indicating the
fraction of outliers: we define the
\code{simple_outlier_fraction} percent of subjects with the highest
absolute values (inverse of pseudo-variance) as outliers.
By default, \code{simple_outlier_fraction=NULL} does not set any outlier
fraction. Either \code{simple_outlier_fraction} or
\code{simple_outlier_thres} should be specified when
\code{robust_method = "simple"}.}

\item{simple_outlier_thres}{a real value, indicating the outliers whose
inverse of pseudo-variance exceed this threshold to be removed from the
current iteration. Either \code{simple_outlier_fraction} or
\code{simple_outlier_thres} should be specified when
\code{robust_method = "simple"}.}

\item{robust_tuning_method}{If \code{robust_tuning_method = "M"},
M-estimation is performed. If \code{robust_tuning_method = "S"},
S-estimation is performed. This argument specifies the tuning method,
i.e., the method for defining outliers in each iteration, when applying
Huber or Bisquare reweighting method (\code{robust_method = "huber"} or
\code{robust_method = "bisquare"}).}
}
\description{
Update each single effect at a time
}
\keyword{internal}
