% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsusie_utils.R
\name{warning_message}
\alias{warning_message}
\title{Utility function to display warning messages as they occur}
\usage{
warning_message(..., style = c("warning", "hint"))
}
\arguments{
\item{...}{warning message}

\item{style}{either "warning" or "hint"}
}
\description{
Utility function to display warning messages as they occur
}
