% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_single_effect_regression.R
\name{weighted_single_effect_regresion}
\alias{weighted_single_effect_regresion}
\title{Weighted single effect regression}
\usage{
weighted_single_effect_regresion(
  y,
  X,
  weights = 1,
  prior_inclusion_prob = NULL,
  coef_prior_variance = 1,
  optimize_V = c("none", "optim", "EM", "simple"),
  check_null_threshold = 0
)
}
\arguments{
\item{weights}{a (n x 1) array, weights of each subject.
For a generalized linear model using iterative reweighted least squared
approach, \eqn{weights = exp(-logw2)}.}

\item{optimize_V}{The optimization method to use for fitting the prior
variance.}

\item{check_null_threshold}{Scalar specifying the threshold on the log-scale
to compare likelihood between current estimate and zero (null).}
}
\value{
A list with the following elements:

\item{alpha}{Vector of posterior inclusion probabilities;
\code{alpha[i]} is posterior probability that the ith
coefficient is non-zero.}

\item{mu}{Vector of posterior means (conditional on inclusion).}

\item{mu2}{Vector of posterior second moments (conditional on inclusion).}

\item{sigma12}{Vector of posterior variance (conditional on inclusion).}

\item{logABF}{Vector of log of asymptotic Bayes factor for each variable.}

\item{logABF_model}{Log of asymptotic Bayes factor for the
weighted single effect regression.}

\item{V}{Prior variance (after optimization if \code{optimize_V != "none"}).}
}
\description{
The WSER function is to compute the posterior distribution of the regression
coefficients of a WSER model. Reference:
\link{https://github.com/stephenslab/susieR/blob/master/R/single_effect_regression.R}
}
\keyword{internal}
