% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_colstats}
\alias{compute_colstats}
\title{Auxilary computation functions}
\usage{
compute_colstats(X, center = FALSE, scale = TRUE)
}
\description{
This file defines the auxilary computation functions.
}
\details{
\code{compute_colstats} sets two attributes for matrix X:
\code{attr(X,'scaled:center')} is a p-vector of column means of X if
\code{center=TRUE}, a p-dim vector of zeros otherwise;
\code{'attr(X,'scaled:scale')} is a p-dim vector of column standard
deviations of X if \code{scale=TRUE}, a p-dim vector of ones otherwise;

\code{compute_Xb} offers the matrix multiplication of
an (n x p) matrix \code{X} and a (p x 1) array/matrix {b}, \ie \eqn{Xb}.
The output is a (n x 1) matrix.
}
\keyword{internal}
