% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_importance_weights.R
\name{robust_importance_weights}
\alias{robust_importance_weights}
\title{Robust estimations}
\usage{
robust_importance_weights(
  values,
  robust_method = c("none", "huber", "simple", "bisquare"),
  simple_outlier_fraction = NULL,
  simple_outlier_thres = NULL,
  robust_tuning_method = c("M", "S"),
  previous_imp_weights = NULL
)
}
\arguments{
\item{values}{a vector that may contains outliers.
May be weights () or residuals(?!)}

\item{robust_method}{If \code{method = "none"}, then all subjects are
included in the coefficient estimation. If \code{method = "simple"},
then the outliers defined by \code{simple_outlier_fraction} or
\code{simple_outlier_thres} are exlucded from the current iteration of
coefficient estimation. If \code{method = "huber"}, huber weights are
assigned to each subject based on the \dQuote{pseudo-residuals} \code{rr}.}

\item{simple_outlier_thres}{a real value, indicating the outliers whose
inverse of pseudo-variance exceed this threshold to be removed from the
current iteration.}

\item{robust_tuning_method}{If \code{robust_tuning_method="M"},
then M-estimation is performed. If \code{robust_tuning_method="S"},
then S-estimation is performed.}

\item{previous_imp_weights}{importance weights in the previous iteration,
used to calculate the importance weights in the current iteration for the
S-estimation. If \param{previous_imp_weights} is NULL, it indicates the
first iteration, and thus S-estimation need to be initialized.}
}
\value{
{imp_weights} An vector of length n, the importance weights.
If a subject is to be removed from the current iteration,
its importance weights is 0.
}
\description{
When fitting a generalized linear model, outliers, i.e. unexpected values,
may occur in both observations and intermediate procedures, such as weights
(i.e. inverse of pseudo-variance) and pseudo-responses. Hence,
robust estimations are be considered to "down-weigh" the influential points.
}
\keyword{internal}
